app/requirements.txt
flask==1.0.2
flask-login
mysql-connector-python==8.0.13